+++ BGL/include/CGAL/boost/graph/Face_filtered_graph.h
@@ -437,7 +437,7 @@ struct Face_filtered_graph
       initialize_halfedge_indices();
   }
 
-  /// changes the set of selected faces using a patch id
+  /// changes the set of selected faces using a patch id.
   template<class FacePatchIndexMap>
   void set_selected_faces(typename boost::property_traits<FacePatchIndexMap>::value_type face_patch_id,
                           FacePatchIndexMap face_patch_index_map)
@@ -506,7 +506,7 @@ struct Face_filtered_graph
     reset_indices();
   }
 
-  /// changes the set of selected faces using a range of face descriptors
+  /// changes the set of selected faces using a range of face descriptors.
   template<class FaceRange>
   void set_selected_faces(const FaceRange& selection)
   {
@@ -570,7 +570,7 @@ struct Face_filtered_graph
     return selected_halfedges[get(himap, halfedge(e,_graph))];
   }
 
-  /// returns the number of selected faces
+  /// returns the number of selected faces.
   size_type number_of_faces() const
   {
     return selected_faces.count();
@@ -659,6 +659,27 @@ struct Face_filtered_graph
     }
 
     return true;
+  }
+
+  /// inverts the selected status of faces.
+  void invert_selection()
+  {
+    selected_faces=~selected_faces;
+    selected_halfedges.reset();
+    selected_vertices.reset();
+
+    for(face_descriptor fd : faces(_graph))
+    {
+      if (!selected_faces.test(get(fimap, fd))) continue;
+      for(halfedge_descriptor hd : halfedges_around_face(halfedge(fd, _graph), _graph))
+      {
+        selected_halfedges.set(get(himap, hd));
+        selected_halfedges.set(get(himap, opposite(hd, _graph)));
+        selected_vertices.set(get(vimap, target(hd, _graph)));
+      }
+    }
+
+    reset_indices();
   }
 
 private: